Properly Sync Birthdays in calendars
This commit fixes issue #31.

Note that when downloading a vCard from a calDav server, we accept
RFC6350 dates of the form YYYYMMDD or --MMDD (the latter is supported
by thunderbird 60, though not by nextcloud 14's contacts app UI. Regardless
nextcloud is happy to store them).  Old fashioned RFC2426 birthdays (BDAY:YYYY-MM-DD) are also supported.

On upload, if the birthday is set or modified, it will be uploaded in RFC6350 format.