Properly Sync Birthdays in calendars (#32)
This commit fixes issue #31.

Note that when downloading a vCard from a calDav server, we have to accept
YYYYMMDD and YYYY-MM-DD formated dates as defined by vCard3 RFC 2426. 

Section 3.1.5 states that the value is a single date value and section 1 says:

   Predefined MIME Directory value specifications used: uri, date,
   date-time, float

These value specifications are in RFC 2425, which says in section 5.8.4 that a date is defined by the ABNF definition:

   date = date-fullyear ["-"] date-month ["-"] date-mday

The ABNF spec is in RFC 2234, section 3.8 of which notes that the square brackets indicate components that are optional.